Update custom_rl_env.py
assign dtype to tensor

+++ custom_rl_env.py
@@ -61,9 +61,9 @@ base_command = {}
 def constant_commands(env: RLTaskEnvCfg) -> torch.Tensor:
     global base_command
     """The generated command from the command generator."""
-    tensor_lst = torch.tensor([0, 0, 0], device=env.device).repeat(env.num_envs, 1)
+    tensor_lst = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float32, device=env.device).repeat(env.num_envs, 1)
     for i in range(env.num_envs):
-        tensor_lst[i] = torch.tensor(base_command[str(i)], device=env.device)
+        tensor_lst[i] = torch.tensor(base_command[str(i)], dtype=torch.float32, device=env.device)
     return tensor_lst
